% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_daily_incidence.R
\name{adjust_daily_incidence}
\alias{adjust_daily_incidence}
\title{Redistribute difference between the original observation and the
imputed incidence value for the single day with reporting anomaly}
\usage{
adjust_daily_incidence(data, adjustment_date, measure)
}
\arguments{
\item{data}{time series data for one location related to an adjustment
case. It has location, date, cum and inc as columns.}

\item{adjustment_date}{date in an adjustment case}

\item{measure}{case or death}
}
\value{
incidence columns after redistribution
}
\description{
Redistribute difference between the original observation and the
imputed incidence value for the single day with reporting anomaly
}
