% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_jhu_global_cum.R
\name{calc_jhu_global_cum}
\alias{calc_jhu_global_cum}
\title{Preprocess JHU global data to get regional counts}
\usage{
calc_jhu_global_cum(jhu_data)
}
\arguments{
\item{jhu_data}{raw JHU global data}
}
\value{
data frame with columns location and date columns
}
\description{
Preprocess JHU global data to get regional counts
}
