% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_jhu_inc.R
\name{calc_jhu_inc}
\alias{calc_jhu_inc}
\title{Assembles a data frame of incident deaths or cases from raw case or death data
from JHU.}
\usage{
calc_jhu_inc(jhu_data)
}
\arguments{
\item{jhu_data}{raw JHU data}
}
\value{
data frame with columns location (fips code), date, and inc
}
\description{
Assembles a data frame of incident deaths or cases from raw case or death data
from JHU.
}
