% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidData-package.R
\docType{package}
\name{covidData}
\alias{covidData}
\alias{covidData-package}
\title{covidData - Versioned data for COVID-19}
\description{
Versioned data for COVID-19
}
\author{
Evan L. Ray \href{mailto:eray@mtholyoke.edu}{eray@mtholyoke.edu}
}
