% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata_download.R
\name{download_healthdata_dailyrevision}
\alias{download_healthdata_dailyrevision}
\title{Download daily revision data at a specific issue date}
\usage{
download_healthdata_dailyrevision(issue_date, healthdata_dailyrevision_history)
}
\arguments{
\item{issue_date}{character issue date (i.e. report date) in format 'yyyy-mm-dd'}

\item{healthdata_dailyrevision_history}{a data.frame with hosp time series data update history}
}
\value{
data.frame with columns state, date, previous_day_admission_adult_covid_confirmed
and previous_day_admission_pediatric_covid_confirmed
}
\description{
Download daily revision data at a specific issue date
}
