% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata_download.R
\name{download_healthdata_timeseries}
\alias{download_healthdata_timeseries}
\title{Download time series data at a specific issue date}
\usage{
download_healthdata_timeseries(issue_date, healthdata_timeseries_history)
}
\arguments{
\item{issue_date}{character issue date (i.e. report date) in format 'yyyy-mm-dd'}

\item{healthdata_timeseries_history}{a data.frame with hosp time series data update history}
}
\value{
data.frame with columns state, date,
previous_day_admission_adult_covid_confirmed
previous_day_admission_pediatric_covid_confirmed, and
previous_day_admission_influenza_confirmed
}
\description{
Download time series data at a specific issue date
}
