% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecdc_locations}
\alias{ecdc_locations}
\title{Available locations in the ECDC Hub}
\format{
A data frame with 32 rows and 3 columns:
\describe{
\item{location_name}{Name of the location}
\item{location}{Location abbreviation}
\item{population}{Location population}
}
}
\usage{
ecdc_locations
}
\description{
Data set with available locations for the EDCD hub
}
\keyword{datasets}
