% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Replace daily inc with NA in specific rows}
\usage{
fill_na(results, adjustments)
}
\arguments{
\item{results}{a result data.frame with location,
date, cum and inc as columns.}

\item{adjustments}{a adjustment case data.frame}
}
\value{
a data.frame with the same set of columns and
NAs on adjustment cases
}
\description{
Replace daily inc with NA in specific rows
}
