% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fips_codes}
\alias{fips_codes}
\title{FIPS codes for the US and its states and territories}
\format{
A data.frame with two-digit FIPS codes and two-letter abbreviations
of states and territories of the US
}
\usage{
fips_codes
}
\description{
FIPS codes for the US and its states and territories
}
\keyword{datasets}
