% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_negative_cases.R
\name{get_negative_cases}
\alias{get_negative_cases}
\title{Find all observations with negative incidence in a
given data frame}
\usage{
get_negative_cases(data)
}
\arguments{
\item{data}{a data frame with location, date, cum and inc}
}
\value{
a data frame of cases with locations and dates
}
\description{
Find all observations with negative incidence in a
given data frame
}
