% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_series_data_link.R
\name{get_time_series_data_link}
\alias{get_time_series_data_link}
\title{Assemble a data frame that stores all links to JHU data series files and
download truth data issued in the most recent week/day.}
\usage{
get_time_series_data_link(
  measure,
  geography = c("US", "global"),
  first_page_only = FALSE,
  download_files = FALSE,
  download_recent = TRUE
)
}
\arguments{
\item{measure}{string specifying measure of covid dynamics:
one of 'deaths' or 'cases'}

\item{geography}{character, which data to read. Default is "US", other option is
"global"}

\item{first_page_only}{boolean specify whether to only scrape
the first page of github repo. Default to FALSE that scrapes all history}

\item{download_files}{boolean specify whether to download truth files after
scraping file links. Default to FALSE}

\item{download_recent}{boolean specify whether to download the most
recent truth file only or all truth files in the most recent week.
Default to TRUE to download the most recent file.}
}
\value{
a data frame with columns date and file_links
}
\description{
Assemble a data frame that stores all links to JHU data series files and
download truth data issued in the most recent week/day.
}
