% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata_download.R
\name{healthdata_dailyrevision_history}
\alias{healthdata_dailyrevision_history}
\title{Retrieve hosp daily revision data update history from healthdata.gov}
\usage{
healthdata_dailyrevision_history()
}
\value{
a data.frame with columns issue_date, date, issue_datetime and file_link
}
\description{
Retrieve hosp daily revision data update history from healthdata.gov
}
