% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{healthdata_hosp_data}
\alias{healthdata_hosp_data}
\title{HealthData.gov data for the US and its states and territories with counts of
hospitalizations}
\format{
A tibble with a row per issue date
}
\usage{
healthdata_hosp_data
}
\description{
HealthData.gov data for the US and its states and territories with counts of
hospitalizations
}
\keyword{datasets}
