% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{healthdata_hosp_early_data}
\alias{healthdata_hosp_early_data}
\title{data frame for hosp time series data before 2021-03-12}
\format{
A data.frame with issue_date and file_link
}
\usage{
healthdata_hosp_early_data
}
\description{
data frame for hosp time series data before 2021-03-12
}
\keyword{datasets}
