% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata_download.R
\name{healthdata_timeseries_history}
\alias{healthdata_timeseries_history}
\title{Retrieve hosp time series data update history from healthdata.gov}
\usage{
healthdata_timeseries_history()
}
\value{
a data.frame with columns issue_date, issue_datetime and file_link
}
\description{
Retrieve hosp time series data update history from healthdata.gov
}
