% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_outliers.R
\name{identify_outliers}
\alias{identify_outliers}
\title{Identify outliers and suitable replacement values}
\usage{
identify_outliers(
  data,
  methods = data.frame(method = c(rep("weekly_extrema_loess_loo", 3)), transform =
    c("none", "sqrt", "log"), ),
  voting_scheme = "majority",
  max_iter = 100,
  ratio_threshold = 3,
  include_outliers_by_method = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{data}{time series data frame for one location with at minimum columns
\code{location}, \code{date} and \code{inc}}
}
\value{
data frame with columns \code{date} and \code{inc} with only rows for
detected outliers.
}
\description{
Identify outliers and suitable replacement values
}
