% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_daily_incidence.R
\name{impute_daily_incidence}
\alias{impute_daily_incidence}
\title{Imputed incidence value for the single day with reporting anomaly}
\usage{
impute_daily_incidence(data, adjustment_date, measure)
}
\arguments{
\item{data}{time series data for one location related to an adjustment
case. It has location, date, cum and inc as columns.}

\item{adjustment_date}{date from adjustment case}

\item{measure}{character vector specifying measure of covid prevalence:
'deaths' or 'cases'}
}
\value{
incidence columns after adjustments
}
\description{
Imputed incidence value for the single day with reporting anomaly
}
