% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu_cases_data}
\alias{jhu_cases_data}
\title{JHU data for the US and its states and territories with counts of cases}
\format{
A tibble with a row per issue date
}
\usage{
jhu_cases_data
}
\description{
JHU data for the US and its states and territories with counts of cases
}
\keyword{datasets}
