% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu_cases_imputed_data}
\alias{jhu_cases_imputed_data}
\title{imputed cases incidence data for negative incidence and adjustment cases}
\format{
A data.frame with location (fips code), date,
measure ('cases') and inc
}
\usage{
jhu_cases_imputed_data
}
\description{
imputed cases incidence data for negative incidence and adjustment cases
}
\keyword{datasets}
