% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu_deaths_data}
\alias{jhu_deaths_data}
\title{JHU data for the US and its states and territories with counts of deaths}
\format{
A tibble with a row per issue date
}
\usage{
jhu_deaths_data
}
\description{
JHU data for the US and its states and territories with counts of deaths
}
\keyword{datasets}
