% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu_deaths_imputed_data}
\alias{jhu_deaths_imputed_data}
\title{Imputed death incidence data for negative incidence and adjustment cases}
\format{
A data.frame with location (fips code), date,
measure ('deaths') and inc
}
\usage{
jhu_deaths_imputed_data
}
\description{
Imputed death incidence data for negative incidence and adjustment cases
}
\keyword{datasets}
