% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu_us_deaths_data_links}
\alias{jhu_us_deaths_data_links}
\title{data frame for JHU US deaths time series data links}
\format{
A data.frame with date and file_link
}
\usage{
jhu_us_deaths_data_links
}
\description{
data frame for JHU US deaths time series data links
}
\keyword{datasets}
