% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_covidcast_data.R
\name{load_covidcast_data}
\alias{load_covidcast_data}
\title{Assemble a data frame of incident hospitalizations due to
COVID-19 as they were available as of a specified issue date in COVIDcast.}
\usage{
load_covidcast_data(
  issue_date = NULL,
  as_of = NULL,
  location_code = NULL,
  spatial_resolution = NULL,
  temporal_resolution = "weekly",
  measure = "hospitalizations",
  geography = "US",
  drop_last_date = FALSE
)
}
\arguments{
\item{issue_date}{character issue date (i.e. report date) to use for
constructing truths in format 'yyyy-mm-dd'}

\item{as_of}{character as_of date to use for constructing truths in
format 'yyyy-mm-dd'}

\item{location_code}{character vector of location codes. Default to NULL.
For US locations, this should be a list of FIPS code or 'US'}

\item{spatial_resolution}{character vector specifying spatial unit types to
include: 'state', 'national' and/or 'county'.
It has to match with locations in \code{location_code}.
Default to NULL, all available spatial resolution based on the given measure.
Note that 'county' is not available for hospitalization measure.}

\item{temporal_resolution}{character vector specifying temporal resolution
to include: 'daily' or 'weekly'}

\item{measure}{character vector specifying measure of covid prevalence:
must be one of 'hospitalizations', 'deaths' or 'cases'.}

\item{geography}{character, which data to read. Default is "US".
Note this variable is not used in the function}

\item{drop_last_date}{boolean indicating whether to drop the last 1 day of
data for the influenza and COVID hospitalization signals. The last day of
data from the HHS data source is unreliable, so it is recommended to set this
to \code{TRUE}. However, the default is \code{FALSE} so that the function maintains
fidelity to the authoritative data source. This argument is ignored if the
\code{measure} is 'deaths' or 'cases'.}
}
\value{
data frame with columns location (fips code), date, inc, and cum
}
\description{
Assemble a data frame of incident hospitalizations due to
COVID-19 as they were available as of a specified issue date in COVIDcast.
}
