% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_data.R
\name{load_jhu_data}
\alias{load_jhu_data}
\title{Assemble a data frame of incident and cumulative deaths or cases due to
COVID-19 as they were available as of a specified issue date.}
\usage{
load_jhu_data(
  issue_date = NULL,
  as_of = NULL,
  location_code = NULL,
  spatial_resolution = "state",
  temporal_resolution = "weekly",
  measure = "deaths",
  geography = c("US", "global")
)
}
\arguments{
\item{issue_date}{character issue date (i.e. report date) to use for
constructing truths in format 'yyyy-mm-dd'}

\item{location_code}{character vector of location codes. Default to NULL.
For US locations, this should be a list of FIPS code or 'US'
For ECDC locations, this should be a list of location name abbreviation.}

\item{spatial_resolution}{character vector specifying spatial unit types to
include: 'county', 'state' and/or 'national'.
This parameter will be ignored if location_code is provided or geography is "global".}

\item{temporal_resolution}{character vector specifying temporal resolution
to include: 'daily' or 'weekly'}

\item{measure}{character vector specifying measure of covid prevalence:
'deaths' or 'cases'}

\item{geography}{character, which data to read. Default is "US", other option is
"global"}
}
\value{
data frame with columns location (fips code), date, inc, and cum
}
\description{
Assemble a data frame of incident and cumulative deaths or cases due to
COVID-19 as they were available as of a specified issue date.
}
