% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata_data.R
\name{preprocess_healthdata_data}
\alias{preprocess_healthdata_data}
\title{Preprocess healthdata data set, calculating incidence, adjusting date, and
calculating national incidence.}
\usage{
preprocess_healthdata_data(raw_healthdata_data, fips_codes)
}
\arguments{
\item{raw_healthdata_data}{tibble one row and columns issue_date and data
The data column should be a list of data frames, with column names
date, state, previous_day_admission_adult_covid_confirmed, and
previous_day_admission_pediatric_covid_confirmed}
}
\value{
a result of similar format to raw_healthdata_data, but columns
date, location, and inc
}
\description{
Preprocess healthdata data set, calculating incidence, adjusting date, and
calculating national incidence.
}
