% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_data.R
\name{preprocess_jhu_data}
\alias{preprocess_jhu_data}
\title{Create a tibble of incident and cumulative deaths or cases based on a
specified issue date and measure. This function will also download a
specified time series data set if issue_date is not available in pre-built
data object.}
\usage{
preprocess_jhu_data(issue_date, measure)
}
\arguments{
\item{issue_date}{character issue date (i.e. report date) to use for
constructing truths in format 'yyyy-mm-dd'}

\item{measure}{character vector specifying measure of covid prevalence:
'deaths' or 'cases'}
}
\value{
tibble with issue_date and data
}
\description{
Create a tibble of incident and cumulative deaths or cases based on a
specified issue date and measure. This function will also download a
specified time series data set if issue_date is not available in pre-built
data object.
}
