% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_negatives.R
\name{replace_negatives}
\alias{replace_negatives}
\title{Replace negative incidence with imputed data and redistribute
residuals across observations before adjustment date}
\usage{
replace_negatives(data, measure)
}
\arguments{
\item{data}{data frame with location, date, cum and inc}

\item{measure}{character vector specifying measure of covid prevalence:
'deaths' or 'cases'}
}
\value{
data frame with location, date, cum and inc after replacement
}
\description{
Replace negative incidence with imputed data and redistribute
residuals across observations before adjustment date
}
