% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_data.R
\name{available_issue_dates}
\alias{available_issue_dates}
\title{Get all available truth data issue dates}
\usage{
available_issue_dates(measure)
}
\arguments{
\item{measure}{character vector specifying measure of covid prevalence:
'deaths', 'cases' or 'hospitalizations'}
}
\value{
date vector of all available issue_date
}
\description{
Get all available truth data issue dates
}
