% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata_download.R
\name{build_healthdata_data}
\alias{build_healthdata_data}
\title{Construct a merged healthdata data set for a single issue date by pulling
from a timeseries update if one is available, or by augmenting the last time
series update on or before the issue date with daily updates that were made
after the last time series update and on or before the issue date}
\usage{
build_healthdata_data(
  issue_date,
  healthdata_timeseries_history,
  healthdata_dailyrevision_history
)
}
\arguments{
\item{issue_date}{issue date}

\item{healthdata_timeseries_history}{data.frame of issue_date and data reported that
date}

\item{healthdata_dailyrevision_history}{data.frame of issue_date, date, and data
reported}
}
\value{
tibble of issue_date and data reported on or by that date
}
\description{
Construct a merged healthdata data set for a single issue date by pulling
from a timeseries update if one is available, or by augmenting the last time
series update on or before the issue date with daily updates that were made
after the last time series update and on or before the issue date
}
