% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_locations}
\alias{global_locations}
\title{Available global locations in JHU}
\format{
A data frame with 190 rows and 2 columns:
\describe{
\item{location_name}{Name of the location}
\item{location}{Location abbreviation}
}
}
\usage{
global_locations
}
\description{
Full list of locations are obtained from CSSEGISandData/COVID-19/csse_covid_19_data
Location abbreviations are from https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2
}
\keyword{datasets}
