% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_data.R
\name{load_jhu_data}
\alias{load_jhu_data}
\title{Construct table of truths for weekly incident and cumulative deaths at
horizons one though 6}
\usage{
load_jhu_data(
  issue_date = NULL,
  spatial_resolution = "state",
  temporal_resolution = "weekly",
  measure = "deaths"
)
}
\arguments{
\item{issue_date}{character issue date (i.e. report date) to use for constructing
truths in format 'yyyy-mm-dd'}

\item{spatial_resolution}{character vector specifying spatial unit types to
include: 'county', 'state' and/or 'national'}

\item{temporal_resolution}{character vector specifying temporal resolution
to include: currently only 'weekly' is supported}

\item{measure}{character vector specifying measure of covid prevalence:
'deaths' or 'cases'}
}
\value{
data frame with
}
\description{
Construct table of truths for weekly incident and cumulative deaths at
horizons one though 6
}
